refactor(message): rehome durable record construction
Keep record identity rules in pure message vocabulary and narrow pane send to delivery mechanics.
